+++ db/migrate/20190323035625_create_courses.rb
@@ -2,8 +2,8 @@ class CreateCourses < ActiveRecord::Migration
   def change
     create_table :courses do |t|
       t.text :name
-      t.decimal :grade
-      t.decimal :avg
+      t.float :grade
+      t.float :avg
       t.references :university, index: true
 
       t.timestamps
